tests: Move billing library tests into test_billing_lib.
Separate the unit tests of corporate/lib/stripe.py -- plan parameter
arithmetic, license ledger, invoicing, support staff actions, and
audit logging -- from the end-to-end Cloud billing flows that remain
in test_stripe.  A few of them use @mock_stripe, so list the module
for full-suite fixture regeneration.
